Move schema related public interfaces to declarative-dsl-api subproject
