+++ platforms/core-configuration/declarative-dsl-api/src/main/java/org/gradle/declarative/dsl/schema/AnalysisSchema.java
@@ -14,10 +14,22 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.declarativedsl.analysis;
+package org.gradle.declarative.dsl.schema;
 
-public interface DataTopLevelFunction extends SchemaFunction {
+import java.io.Serializable;
+import java.util.Map;
+import java.util.Set;
 
-    String getPackageName();
+public interface AnalysisSchema extends Serializable {
+
+    DataClass getTopLevelReceiverType();
+
+    Map<FqName, DataClass> getDataClassesByFqName();
+
+    Map<FqName, DataTopLevelFunction> getExternalFunctionsByFqName();
+
+    Map<FqName, ExternalObjectProviderKey> getExternalObjectsByFqName();
+
+    Set<FqName> getDefaultImports();
 
 }
