+++ platforms/core-configuration/declarative-dsl-api/src/main/java/org/gradle/declarative/dsl/schema/DataTypeRef.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.declarativedsl.analysis;
+package org.gradle.declarative.dsl.schema;
 
 import java.io.Serializable;
